UPSTREAM: ALSA: rawmidi: Add framing mode
This commit adds a new framing mode that frames all MIDI data into
32-byte frames with a timestamp.

The main benefit is that we can get accurate timestamps even if
userspace wakeup and processing is not immediate.

Testing on a Celeron N3150 with this mode has a max jitter of 2.8 ms,
compared to the in-kernel seq implementation which has a max jitter
of 5 ms during idle and much worse when running scheduler stress tests
in parallel.

Signed-off-by: David Henningsson <coding@diwic.se>
Reviewed-by: Jaroslav Kysela <perex@perex.cz>
Link: https://lore.kernel.org/r/20210515071533.55332-1-coding@diwic.se
Signed-off-by: Takashi Iwai <tiwai@suse.de>
(cherry picked from commit 08fdced60ca08e34e316a3ab945636fcdfcbc973)

BUG=b:238403794
TEST=Test Audio use cases.

Signed-off-by: Rutumber Nath <rutumberx.nath@intel.com>
Change-Id: I675345b9a3962b5cc98eb14af062484fab800fb8